Add the settings modal
