+++ src/services/DiscordService.ts
@@ -2,15 +2,15 @@ import axios from "axios";
 import { useDispatch, useSelector } from "react-redux";
 
 import { useEffect } from "react";
-import { getAccessToken, getTokenType } from "../redux/auth/authSelectors";
+import { AuthSelectors } from "../redux/auth/authSelectors";
 import { UserAction } from "../redux/user/userActions";
 import { UserSelectors } from "../redux/user/userSelectors";
 
 const useCurrentUserInfo = () => {
     const dispatch = useDispatch();
 
-    const authTokenType = useSelector(getTokenType);
-    const accessToken = useSelector(getAccessToken);
+    const authTokenType = useSelector(AuthSelectors.getTokenType);
+    const accessToken = useSelector(AuthSelectors.getAccessToken);
     const isUserSignedIn = useSelector(UserSelectors.getUsername) !== undefined;
 
     const endpoint = "https://discord.com/api/users/@me";
